fixing bug

+++ dev_ws/src/field_robot/src/navigation/bt_plugins/action/store_turn_information_node.cpp
@@ -27,7 +27,7 @@ class StoreTurnInformation : public nav2_behavior_tree::BtActionNode<field_robot
         static BT::PortsList providedPorts()
         {
           return providedBasicPorts({
-            BT::OutputPort<int8>("remaining_rows"),
+            BT::OutputPort<int>("remaining_rows"),
             BT::OutputPort<geometry_msgs::msg::PoseStamped>("last_ron_pose"),
             BT::OutputPort<bool>("turn_direction"),
           });
